Scripts/requirements.txt
kafka-python
influxdb-client
pandas